% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welcome.R
\name{Welcome}
\alias{Welcome}
\title{FUNCTION_TITLE}
\usage{
Welcome(n)
}
\arguments{
\item{n}{a counter.}
}
\value{
a message.
}
\description{
A hello, world extension.
}
\details{
It simply says times in the message.
}
\keyword{internal}
