% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabates.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Dataset}
\format{
A data frame with 1,000 rows and 14 variables:
\describe{
\item{\code{ID}}{ID.}
\item{\code{No_Pation}}{No. of Patient.}
\item{\code{Gender}}{Gender.}
\item{\code{AGE}}{Age.}
\item{\code{Urea}}{Urea.}
\item{\code{Cr}}{Creatinine ratio(Cr).}
\item{\code{HbA1c}}{HbA1c.}
\item{\code{Chol}}{Cholesterol (Chol).}
\item{\code{TG}}{Triglycerides(TG).}
\item{\code{HDL}}{HDL.}
\item{\code{LDL}}{LDL.}
\item{\code{VLDL}}{VLDL.}
\item{\code{BMI}}{Body Mass Index (BMI).}
\item{\code{CLASS}}{Class (the patient's diabetes disease class may be Diabetic, Non-Diabetic, or Predict-Diabetic).}
}
}
\source{
Rashid A (2020), “Diabetes Dataset”, Mendeley Data, V1, doi: 10.17632/wj9rwkp9c2.1.
}
\usage{
diabetes
}
\description{
A diabetes dataset on 1,000 patients.
}
\details{
The data were collected from the Iraqi society, as they data were acquired from the laboratory of Medical City Hospital
and (the Specializes Center for Endocrinology and Diabetes-Al-Kindy Teaching Hospital).
}
\examples{
\dontrun{
data(diabetes)
head(diabetes)
}
}
\seealso{
\code{\link[gaawr2]{DiaHealth}}.
}
\keyword{datasets}
\keyword{internal}
