% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaawr2.R
\docType{package}
\name{gaawr2}
\alias{gaawr2}
\alias{gaawr2-package}
\title{Genetic Association Analysis with R (II)}
\description{
This package illustrates creation and maintenance of package used in genetic association analysis.
}
\details{
Available data are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{Datasets} \tab   \cr
   \code{\link{biomaRt}} \tab Curated data from biomaRt \cr
}


Some generic description for the datasets are as follows.
\itemize{
\item chr Chromosome.
\item start Start position.
\item end End position.
\item gene Gene name.
\item UniProt UniProt ID.
}
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item An Overview of pQTLdata. \code{vignette("gaawr2")}.
}
}

\examples{
\dontrun{
# See more on SCALLOP-INF at https://jinghuazhao.github.io/INF/
# datasets
head(biomaRt)

}
}
\author{
Jing Hua Zhao in collaboration with other colleagues.
}
\keyword{internal}
