% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaawr2.R
\docType{package}
\name{gaawr2}
\alias{gaawr2}
\alias{gaawr2-package}
\title{Genetic Association Analysis with R (II)}
\description{
An R package companion to the titled Henry-Stewart talk, part 2.
}
\details{
Available data are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{Functions} \tab   \cr
   \code{\link{welcome}} \tab An enhanced welcome \cr
}
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item Genetic Association Analysis with R (II), \code{vignette("gaawr2")}.
\item Web facilities, \code{vignette("web")}.
}
}

\examples{
welcome(3)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jinghuazhao.github.io/gaawr2/}
  \item \url{https://github.com/jinghuazhao/gaawr2}
  \item Report bugs at \url{https://github.com/jinghuazhao/gaawr2/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues.
}
\keyword{internal}
