% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaawr2.R
\docType{package}
\name{gaawr2}
\alias{gaawr2}
\alias{gaawr2-package}
\title{Genetic Association Analysis (II)}
\description{
It renders details to Henry-Stewart talks by Zhao (2009, \doi{10.69645/DCRY5578}), which showcases analysis in aspects such as testing of polymorphic variant(s) for Hardy-Weinberg equilibrium, association with trait using genetic and statistical models as well as Bayesian implementation, power calculation in study design and genetic annotation. It also covers R integration with the Linux environment, GitHub, package creation and web applications.
}
\details{
Available data and function are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{Dataset} \tab   \cr
   \code{\link{DiaHealth}} \tab A Bangladeshi dataset for Type 2 diabetes prediction \cr
   \code{\link{diabetes}} \tab A diabetes dataset \cr
   \strong{Functions} \tab   \cr
   \code{\link{welcome}} \tab An enhanced welcome \cr
}


We can add references such as \insertCite{Rpack:bibtex;textual}{Rdpack}.
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item Genetic Association Analysis with R (II), \code{vignette("gaawr2")}.
\item Web facilities, \code{vignette("web")}.
}
}

\examples{
welcome(3)
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jinghuazhao.github.io/gaawr2/}
  \item \url{https://github.com/jinghuazhao/gaawr2}
  \item Report bugs at \url{https://github.com/jinghuazhao/gaawr2/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues.
}
\keyword{internal}
