% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.R
\name{welcome}
\alias{welcome}
\title{welcome function}
\usage{
welcome(n)
}
\arguments{
\item{n}{The number of times (>1 integer) to welcome the user.}
}
\value{
Prints a welcome message to the console.
}
\description{
It prints a welcome message, saying number of times.
}
\examples{
welcome(3)

}
